Code fixes for clearning the digit buffer, code refactoring to provide the ILibrary interface within the SIP Plugin so that modification to the LineManager is not needed.
