+++ ivrToolkit.Core/ivrToolkit.Core.csproj
@@ -85,7 +85,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AbstractScript.cs" />
-    <Compile Include="ILibrary.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
       <AutoGen>True</AutoGen>
