+++ ivrToolkit.DialogicSipPluginSync/ILibrary.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace ivrToolkit.Core
+namespace ivrToolkit.DialogicSipPluginSync
 {
     public interface ILibrary
     {
