+++ ivrToolkit.sln
@@ -104,6 +104,7 @@ Global
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Debug|x86.Build.0 = Debug|Any CPU
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Release|Any CPU.Build.0 = Release|Any CPU
 		{760FE3FE-9718-4A47-AC0E-8820818F1B15}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -116,6 +117,7 @@ Global
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Debug|x86.Build.0 = Debug|Any CPU
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3AC20DB2-78EB-4F85-BDE8-FA0C58F44E51}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -128,6 +130,7 @@ Global
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Debug|x86.Build.0 = Debug|Any CPU
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B98E1085-F4C1-40FE-9EE5-2BC8EDA3B3A3}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -156,6 +159,7 @@ Global
 		{61C95321-283C-4D06-869E-141F41C3F679}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{61C95321-283C-4D06-869E-141F41C3F679}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{61C95321-283C-4D06-869E-141F41C3F679}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{61C95321-283C-4D06-869E-141F41C3F679}.Debug|x86.Build.0 = Debug|Any CPU
 		{61C95321-283C-4D06-869E-141F41C3F679}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{61C95321-283C-4D06-869E-141F41C3F679}.Release|Any CPU.Build.0 = Release|Any CPU
 		{61C95321-283C-4D06-869E-141F41C3F679}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -168,6 +172,7 @@ Global
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Debug|x86.Build.0 = Debug|Any CPU
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Release|Any CPU.Build.0 = Release|Any CPU
 		{77B6E071-41AA-4D65-B043-68421FFD2C96}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -180,6 +185,7 @@ Global
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Debug|x86.Build.0 = Debug|Any CPU
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FDA6EFE3-013E-4F5D-AB19-3009D87324C2}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -192,6 +198,7 @@ Global
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Debug|x86.Build.0 = Debug|Any CPU
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8261B4EC-8999-4805-BBB0-EBEF64B4701F}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -204,6 +211,7 @@ Global
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Debug|x86.Build.0 = Debug|Any CPU
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3E08C1E3-7D2C-4D13-8168-E08234752DFB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -260,6 +268,7 @@ Global
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Debug|x86.Build.0 = Debug|Any CPU
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E3D0210D-D86E-42B2-9773-673BBE6678B3}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
@@ -272,6 +281,7 @@ Global
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Debug|Win32.ActiveCfg = Debug|Any CPU
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Debug|x86.Build.0 = Debug|Any CPU
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{570A7BE4-8BF3-4B76-B37F-85A0C8576B2F}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
